Add files via upload
Assignment_05 tasked each student to create a python script that allows a user to modify and save a to do list of items and their corresponding priorities. The items should be saved in a table (a list that contains dictionary rows with “Tasks” and Priorities”). The user should be able to view the list, add an item, remove and item, save the list, as well as save and exit the program.